backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.16
psycopg2-binary==2.9.9
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.2
pydantic-settings==2.0.3
passlib[bcrypt]==1.7.4